economic-data-platform/pyproject.toml
[tool.poetry]
name = "economic-data-platform"
version = "0.1.0"
description = "Multi-domain data analytics platform"
authors = ["Data Team <data@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
numpy = "^1.26.3"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pymssql = "^2.2.11"
clickhouse-driver = "^0.2.6"
redis = "^5.0.1"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
requests = "^2.31.0"
httpx = "^0.26.0"
apache-airflow = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
black = "^24.1.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
